fix(lexmode): expand LexState to uint32 to fix large grammar truncation
- Change LexState and AfterWhitespaceLexState from uint16 to uint32 in LexMode struct
- Large grammars (Fortran, COBOL) exceed 65k DFA states combined across modes; uint16 truncation silently corrupted lex mode pointers causing parse failures
- Add VerifyLexModeCompleteness diagnostic to detect when action table expects terminals unreachable from current lex mode
- Add Fortran parity tests covering integer kind suffixes, parametric types, and array subscripts
- Add GOT_DEBUG_LEXMODE_BUILD and GOT_DEBUG_LEXMODE_COMPUTE env flags for lex mode diagnostics
- Update go.mod: add stretchr/testify dependency, remove unused LSP protocol deps

+++ cmd/ts2go/embedded.go
@@ -94,7 +94,7 @@ func BuildLanguage(g *ExtractedGrammar) *gotreesitter.Language {
 		lang.LexModes = make([]gotreesitter.LexMode, len(g.LexModes))
 		for i, lm := range g.LexModes {
 			lang.LexModes[i] = gotreesitter.LexMode{
-				LexState:          uint16(lm.LexState),
+				LexState:          uint32(lm.LexState),
 				ExternalLexState:  uint16(lm.ExternalLexState),
 				ReservedWordSetID: uint16(lm.ReservedWordSetID),
 			}
